Fixed small for loop bug. Re-organized project's includes
