+++ source/my_includes/cameraClass.h
